% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catastrophic.R
\name{as_costs}
\alias{as_costs}
\title{Formulate survey data to a cost object}
\usage{
as_costs(dat, inc_guard = T, explt.fn = "Median",
  adjust_anchor = "G3")
}
\arguments{
\item{dat}{source data}

\item{inc_guard}{include guardian cost or not}

\item{explt.fn}{function for expolation if data missed, median as default}

\item{adjust_anchor}{the key layer of cost structure}
}
\description{
Formulate survey data to a cost object
}
\examples{

}
