% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cata_timing.R
\name{calc_time_to_cc}
\alias{calc_time_to_cc}
\alias{summary.TimeToCataCosts}
\alias{print.summaryTimeToCataCosts}
\title{Calculate Time to Catastrophic Costs}
\usage{
calc_time_to_cc(cs, sp = c("none", "all spe", "all sen", "prop",
  "explicit"), sp_sen = 1, indirect = c("hc", "dinc", "prop"),
  indirect_hc = 1, threshold = 0.2)

\method{summary}{TimeToCataCosts}(ttcc)

\method{print}{summaryTimeToCataCosts}(res)
}
\arguments{
\item{cs}{an object of Costs}

\item{sp}{option for specifying social protection}

\item{sp_sen}{proportion of social protection are TB-sensitive}

\item{indirect}{option for specifying indiriect cost}

\item{indirect_hc}{weight of indiriect cost in human capital}

\item{threshold}{threshold of catastrophic costs}
}
\value{
an object of TimeToCataCosts
}
\description{
Calculate Time to Catastrophic Costs
}
