% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cc.R
\name{vis_cc_age}
\alias{vis_cc_age}
\title{Visualise catastropoic patient costs by age and sex}
\usage{
vis_cc_age(catas, by = 50)
}
\arguments{
\item{catas}{A CataCosts object}

\item{by}{size of breaks of population size}
}
\value{
age pyramid of cases with catastrophic patient costs
}
\description{
Visualise catastropoic patient costs by age and sex
}
\examples{
Costs <- as_costs(Survey)
Catas <- calc_cc(Costs, sp = "all spe", indirect = "hc")

vis_cc_age(Catas)
vis_cc_age_mdr(Catas)

}
