% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cost.R
\name{vis_type_age}
\alias{vis_type_age}
\title{Title}
\usage{
vis_type_age(catas, scale = 1000,
  xlab = "Costs per patient (NTD in thousand)")
}
\arguments{
\item{catas}{}

\item{scale}{}

\item{xlab}{}
}
\description{
Title
}
