% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catastrophic.R
\name{calc_cc}
\alias{calc_cc}
\title{Title}
\usage{
calc_cc(cs, sp = c("none", "all spe", "all sen", "prop", "explicit"),
  sp_sen = 1, indirect = c("hc", "dinc", "prop"), indirect_hc = 1,
  hhi = c("annual", "treatment"), threshold = 0.2)
}
\arguments{
\item{cs}{}

\item{sp}{}

\item{sp_sen}{}

\item{indirect}{}

\item{indirect_hc}{}

\item{hhi}{}

\item{threshold}{}
}
\description{
Title
}
